% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{color_heatmap}
\alias{color_heatmap}
\title{color_heatmap}
\usage{
color_heatmap(sample_metadata)
}
\arguments{
\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}
}
\description{
Helper function to color the heatmap
}
