% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extract_kegg_terms}
\alias{extract_kegg_terms}
\title{extract_kegg_terms}
\usage{
extract_kegg_terms(gene, species)
}
\arguments{
\item{gene}{A character vector containing the Entrez-IDs.}

\item{species}{A character string indicating the sample's species.
Currently supported: "Hs".}
}
\description{
Helper function to extract KEGG terms
}
