% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{go_enrichment_analysis}
\alias{go_enrichment_analysis}
\title{go_enrichment_analysis}
\usage{
go_enrichment_analysis(
  gene,
  org_db,
  organism,
  fitlm,
  pathwayid,
  species,
  plotdir = NULL
)
}
\arguments{
\item{gene}{A character vector containing the Entrez-IDs.}

\item{org_db}{org_db}

\item{organism}{A character string indicating the organism.
Currently supported: "hsa".}

\item{fitlm}{An MArrayLM object, returned by `go_diff_reg()`.}

\item{pathwayid}{KEGG pathway ID (like hsa04110 for cell cycle)}

\item{species}{A character string indicating the sample's species.
Currently supported: "Hs".}

\item{plotdir}{A character string. Path to the folder to store resulting
plots. Default: "./plots/".}
}
\description{
Helper function to perform enrichment analysis
}
