% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_deg_heatmap}
\alias{plot_deg_heatmap}
\title{plot_deg_heatmap}
\usage{
plot_deg_heatmap(mergeset, genes, patientcolors, filename = NULL)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{genes}{A object. The output of the function `identify_degs()`.}

\item{patientcolors}{A object. The ouput of the function `color_heatmap()`.}

\item{filename}{A character string indicating the filename. If default
(\code{NULL}) a plot named `DEG_heatmap.png` will be created
inside the directory "./plots".}
}
\description{
Helper function to create DEG heatmap
}
