% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_enriched_barplot}
\alias{plot_enriched_barplot}
\title{plot_enriched_barplot}
\usage{
plot_enriched_barplot(enrichmentobj, type, filename = NULL, show_category = 20)
}
\arguments{
\item{enrichmentobj}{An object resulting from the function
`go_enrichment_analysis()`.}

\item{type}{A character string. One of eiter "GO" or "KEGG".}

\item{filename}{A character string indicating the filename. If default
(\code{NULL}) a plot named `Enriched_{type}.png` will be created
inside the directory "./plots".}

\item{show_category}{An integer. Indicating the number of maximum categories
to show in barplot.}
}
\description{
Helper function to create enrichted barplots
}
