% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigidentDEG.R
\name{sigidentDEG}
\alias{sigidentDEG}
\title{Perform DEG Analysis in Gene Expression Datasets Derived
  from MicroArrays}
\usage{
sigidentDEG(
  mergeset,
  mergedset,
  sample_metadata,
  diagnosis,
  idtype,
  fdr,
  plotdir = "./plots/",
  csvdir = "./tables/"
)
}
\arguments{
\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{mergedset}{A large merged Expression Data set. The output
of the funtion `sigident.preproc::load_geo_data()`.}

\item{sample_metadata}{A data frame. The data frame holding the
sample metadata.}

\item{diagnosis}{A vector of integers, holding the binary outcome variable
(0 = "Control", 1 = "Target").}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{fdr}{A positive numeric value between (max. 0.05) indicating the
desired q-Value during DEG analysis (Default: 0.01).}

\item{plotdir}{A character string. Path to the folder to store resulting
plots. Default: "./plots/".}

\item{csvdir}{A character string. Path to the folder to store output
tables. Default: "./tables/".}
}
\description{
One function to perform DEG analysis.
}
