% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigidentEnrichment.R
\name{sigidentEnrichment}
\alias{sigidentEnrichment}
\title{Perform Enrichtment Analysis in Gene Expression Datasets
  Derived from MicroArrays}
\usage{
sigidentEnrichment(
  mergedset,
  mergeset,
  idtype,
  diagnosis,
  species,
  org_db,
  organism,
  pathwayid,
  plotdir = "./plots/",
  csvdir = "./tables/"
)
}
\arguments{
\item{mergedset}{A large merged Expression Data set. The output
of the funtion `sigident.preproc::load_geo_data()`.}

\item{mergeset}{A matrix of merged expression sets (rows = genes,
columns = samples). The output of the funtion
`sigident.preproc::load_geo_data()`.}

\item{idtype}{A character string. The type of ID used to name the
genes. One of 'entrez' or 'affy' intended to use either entrez IDs or
affy IDs. Caution: when using entrez IDs, missing and duplicated IDs
are being removed!}

\item{diagnosis}{A vector of integers, holding the binary outcome variable
(0 = "Control", 1 = "Target").}

\item{species}{A character string indicating the sample's species.
Currently supported: "Hs".}

\item{org_db}{A character string indicating the OrgDb. Currently
supported: "org.Hs.eg.db".}

\item{organism}{A character string indicating the organism.
Currently supported: "hsa".}

\item{pathwayid}{A character string indicating the pathway to show
in the enrichment analysis. Currently supported: "hsa04110".'}

\item{plotdir}{A character string. Path to the folder to store resulting
plots. Default: "./plots/".}

\item{csvdir}{A character string. Path to the folder to store output
tables. Default: "./tables/".}
}
\description{
One function to perform gene enrichment analysis.
}
